+++ packages/ai/core/generate-text/stream-text-result.ts
@@ -1,10 +1,6 @@
 import { ServerResponse } from 'node:http';
-import {
-  AIStreamCallbacksAndOptions,
-  CoreAssistantMessage,
-  CoreToolMessage,
-  StreamData,
-} from '../../streams';
+import { StreamData } from '../../streams/stream-data';
+import { CoreAssistantMessage, CoreToolMessage } from '../prompt/message';
 import { CoreTool } from '../tool';
 import {
   CallWarning,
@@ -13,10 +9,7 @@ import {
   LogProbs,
   ProviderMetadata,
 } from '../types';
-import {
-  LanguageModelResponseMetadata,
-  LanguageModelResponseMetadataWithHeaders,
-} from '../types/language-model-response-metadata';
+import { LanguageModelResponseMetadata } from '../types/language-model-response-metadata';
 import { LanguageModelUsage } from '../types/usage';
 import { AsyncIterableStream } from '../util/async-iterable-stream';
 import { StepResult } from './step-result';
@@ -141,21 +134,6 @@ need to be added separately.
   readonly fullStream: AsyncIterableStream<TextStreamPart<TOOLS>>;
 
   /**
-  Converts the result to an `AIStream` object that is compatible with `StreamingTextResponse`.
-  It can be used with the `useChat` and `useCompletion` hooks.
-
-  @param callbacks
-  Stream callbacks that will be called when the stream emits events.
-
-  @returns A data stream.
-
-  @deprecated Use `toDataStream` instead.
-     */
-  toAIStream(
-    callbacks?: AIStreamCallbacksAndOptions,
-  ): ReadableStream<Uint8Array>;
-
-  /**
   Converts the result to a data stream.
 
   @param data an optional StreamData object that will be merged into the stream.
@@ -169,21 +147,6 @@ need to be added separately.
     getErrorMessage?: (error: unknown) => string;
     sendUsage?: boolean; // default to true (change to false in v4: secure by default)
   }): ReadableStream<Uint8Array>;
-
-  /**
-  Writes stream data output to a Node.js response-like object.
-  It sets a `Content-Type` header to `text/plain; charset=utf-8` and
-  writes each stream data part as a separate chunk.
-
-  @param response A Node.js response-like object (ServerResponse).
-  @param init Optional headers and status code.
-
-  @deprecated Use `pipeDataStreamToResponse` instead.
-     */
-  pipeAIStreamToResponse(
-    response: ServerResponse,
-    init?: { headers?: Record<string, string>; status?: number },
-  ): void;
 
   /**
   Writes data stream output to a Node.js response-like object.
@@ -213,21 +176,6 @@ need to be added separately.
   @param init Optional headers, status code, and status text.
      */
   pipeTextStreamToResponse(response: ServerResponse, init?: ResponseInit): void;
-
-  /**
-  Converts the result to a streamed response object with a stream data part stream.
-  It can be used with the `useChat` and `useCompletion` hooks.
-
-  @param options An object with an init property (ResponseInit) and a data property.
-  You can also pass in a ResponseInit directly (deprecated).
-
-  @return A response object.
-
-  @deprecated Use `toDataStreamResponse` instead.
-     */
-  toAIStreamResponse(
-    options?: ResponseInit | { init?: ResponseInit; data?: StreamData },
-  ): Response;
 
   /**
   Converts the result to a streamed response object with a stream data part stream.