+++ packages/react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/react
 
+## 1.0.0-canary.4
+
+### Major Changes
+
+- ca3e586: chore (ui): remove experimental_useAssistant export
+
+### Patch Changes
+
+- Updated dependencies [dce4158]
+- Updated dependencies [dce4158]
+  - @ai-sdk/provider-utils@2.0.0-canary.2
+  - @ai-sdk/ui-utils@1.0.0-canary.4
+
 ## 1.0.0-canary.3
 
 ### Major Changes